+++ aem-intellij-lang/src/main/kotlin/com/aemtools/lang/jcrproperty/editor/JpBraceMatcher.kt
@@ -13,7 +13,6 @@ class JpBraceMatcher : PairedBraceMatcher {
   override fun getCodeConstructStart(file: PsiFile?, openingBraceOffset: Int): Int = openingBraceOffset
 
   override fun getPairs(): Array<BracePair> = arrayOf(
-      BracePair(JpTypes.LBRACE, JpTypes.RBRACE, false),
       BracePair(JpTypes.LBRACKET, JpTypes.RBRACKET, false)
   )