+++ aem-intellij-lang/src/main/kotlin/com/aemtools/lang/jcrproperty/highlight/JpHighlighter.kt
@@ -1,24 +1,7 @@
 package com.aemtools.lang.jcrproperty.highlight
 
 import com.aemtools.lang.jcrproperty.colorscheme.JpColors
-import com.aemtools.lang.jcrproperty.psi.JpTypes.BINARY
-import com.aemtools.lang.jcrproperty.psi.JpTypes.BOOLEAN
-import com.aemtools.lang.jcrproperty.psi.JpTypes.COMMA
-import com.aemtools.lang.jcrproperty.psi.JpTypes.DATE
-import com.aemtools.lang.jcrproperty.psi.JpTypes.DECIMAL
-import com.aemtools.lang.jcrproperty.psi.JpTypes.DOUBLE
-import com.aemtools.lang.jcrproperty.psi.JpTypes.LBRACE
-import com.aemtools.lang.jcrproperty.psi.JpTypes.LBRACKET
-import com.aemtools.lang.jcrproperty.psi.JpTypes.LONG
-import com.aemtools.lang.jcrproperty.psi.JpTypes.NAME
-import com.aemtools.lang.jcrproperty.psi.JpTypes.PATH
-import com.aemtools.lang.jcrproperty.psi.JpTypes.RBRACE
-import com.aemtools.lang.jcrproperty.psi.JpTypes.RBRACKET
-import com.aemtools.lang.jcrproperty.psi.JpTypes.REFERENCE
-import com.aemtools.lang.jcrproperty.psi.JpTypes.STRING
-import com.aemtools.lang.jcrproperty.psi.JpTypes.TYPE
-import com.aemtools.lang.jcrproperty.psi.JpTypes.URI
-import com.aemtools.lang.jcrproperty.psi.JpTypes.WEAK_REFERENCE
+import com.aemtools.lang.jcrproperty.psi.JpTypes.*
 import com.intellij.lexer.Lexer
 import com.intellij.openapi.editor.colors.TextAttributesKey
 import com.intellij.openapi.fileTypes.SyntaxHighlighterBase
@@ -36,12 +19,14 @@ class JpHighlighter : SyntaxHighlighterBase() {
 
   private fun map(tokenType: IElementType?) = when (tokenType) {
     LBRACE,
-    RBRACE -> JpColors.DELIMITER
+    RBRACE -> JpColors.BRACES
 
     LBRACKET,
-    RBRACKET -> JpColors.BRACES
+    RBRACKET -> JpColors.BRACKETS
 
-    COMMA -> JpColors.TYPE
+    COMMA -> JpColors.DELIMITER
+
+    SINGLE, ARRAY_VALUE -> JpColors.VALUE
 
     BINARY, BOOLEAN,
     DATE, DECIMAL,